requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
pymongo==4.3.3
python-dotenv==0.21.1
pydantic==1.10.4
